+++ dflib/src/main/java/org/dflib/exp/num/DoubleExpFactory.java
@@ -128,6 +128,11 @@ public class DoubleExpFactory extends NumericExpFactory {
     }
 
     @Override
+    public NumExp<Long> round(Exp<? extends Number> exp) {
+        return LongExp1.mapVal("round", cast(exp), n -> Math.round(n));
+    }
+
+    @Override
     public Condition eq(Exp<? extends Number> left, Exp<? extends Number> right) {
         return DoubleCondition2.mapVal("=", cast(left), cast(right), Double::equals, DoubleSeries::eq);
     }